+++ hit.h
@@ -18,7 +18,7 @@ class Object;
 class Hit {
 public:
 	bool flag;
-	double t;
+	long double t;
 	Object *what;
 	Vertex position;
 	Vector normal;
